server/requirements.txt
flask==3.0.3
flask-cors==5.0.0
firebase-admin==6.5.0
pydantic==2.8.2

scikit-learn==1.5.2
xgboost==2.1.2
prophet==1.1.6
pandas==2.2.3
numpy==2.1.3
joblib==1.4.2
scipy==1.14.1

google-cloud-storage==2.18.2
google-cloud-firestore==2.19.0
google-api-python-client==2.151.0
google-api-core==2.22.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1

gunicorn==22.0.0
cmdstanpy==1.2.4
